// src/plugins/product-types/ui/generated-admin-types.ts
export type Maybe<T> = T;
export type InputMaybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Money: { input: number; output: number; }
  Upload: { input: any; output: any; }
};
// ...
export type Administrator = Node & {
  __typename?: 'Administrator';
  createdAt: Scalars['DateTime']['output'];
  customFields?: Maybe<Scalars['JSON']['output']>;
  emailAddress: Scalars['String']['output'];
  firstName: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  lastName: Scalars['String']['output'];
  updatedAt: Scalars['DateTime']['output'];
  user: User;
};

export type AdministratorFilterParameter = {
  createdAt?: InputMaybe<DateOperators>;
  emailAddress?: InputMaybe<StringOperators>;
  firstName?: InputMaybe<StringOperators>;
  id?: InputMaybe<IdOperators>;
  lastName?: InputMaybe<StringOperators>;
  updatedAt?: InputMaybe<DateOperators>;
};

export type AdministratorList = PaginatedList & {
  __typename?: 'AdministratorList';
  items: Array<Administrator>;
  totalItems: Scalars['Int']['output'];
};
// ...
export type Asset = Node & {
  __typename?: 'Asset';
  createdAt: Scalars['DateTime']['output'];
  customFields?: Maybe<Scalars['JSON']['output']>;
  fileSize: Scalars['Int']['output'];
  focalPoint?: Maybe<Coordinate>;
  height: Scalars['Int']['output'];
  id: Scalars['ID']['output'];
  mimeType: Scalars['String']['output'];
  name: Scalars['String']['output'];
  preview: Scalars['String']['output'];
  source: Scalars['String']['output'];
  tags: Array<Tag>;
  type: AssetType;
  updatedAt: Scalars['DateTime']['output'];
  width: Scalars['Int']['output'];
};
// ...
export type CreateProductTypeInput = {
  additionalType: Scalars['String']['input'];
  alternateName: Scalars['String']['input'];
  disambiguatingDescription: Scalars['String']['input'];
  durationUnit?: InputMaybe<DurationUnit>;
  isService: Scalars['Boolean']['input'];
  meta_description: Scalars['String']['input'];
  meta_keyword: Scalars['String']['input'];
  meta_title: Scalars['String']['input'];
  oneWordName: Scalars['String']['input'];
  productId: Scalars['ID']['input'];
  repetitions: Array<InputMaybe<Scalars['Int']['input']>>;
  serviceOutput?: InputMaybe<Scalars['String']['input']>;
  serviceType?: InputMaybe<Scalars['String']['input']>;
  slogan: Scalars['String']['input'];
  transmissionDays?: InputMaybe<Scalars['Int']['input']>;
};
// ...
export enum DurationUnit {
  Days = 'Days',
  Months = 'Months',
  Weeks = 'Weeks'
}
// ...
export type OrderMessageList = PaginatedList & {
  __typename?: 'OrderMessageList';
  items: Array<OrderMessage>;
  totalItems: Scalars['Int']['output'];
};
// ...
export type ProductType = Node & {
  __typename?: 'ProductType';
  additionalType: Scalars['String']['output'];
  alternateName: Scalars['String']['output'];
  disambiguatingDescription: Scalars['String']['output'];
  durationUnit?: Maybe<DurationUnit>;
  id: Scalars['ID']['output'];
  isService: Scalars['Boolean']['output'];
  meta_description: Scalars['String']['output'];
  meta_keyword: Scalars['String']['output'];
  meta_title: Scalars['String']['output'];
  oneWordName: Scalars['String']['output'];
  productId: Scalars['ID']['output'];
  repetitions: Array<Maybe<Scalars['Int']['output']>>;
  serviceOutput?: Maybe<Scalars['String']['output']>;
  serviceType?: Maybe<Scalars['String']['output']>;
  slogan: Scalars['String']['output'];
  transmissionDays?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Query = {
  __typename?: 'Query';
  activeAdministrator?: Maybe<Administrator>;
  activeChannel: Channel;
  administrator?: Maybe<Administrator>;
  administrators: AdministratorList;
  allOrderMessages: OrderMessageList;
  /** Get a single Asset by id */
  asset?: Maybe<Asset>;
  /** Get a list of Assets */
  assets: AssetList;
  availableOrderExportStrategies: Array<Scalars['String']['output']>;
  campaign?: Maybe<Campaign>;
  campaigns: CampaignList;
  channel?: Maybe<Channel>;
  channels: ChannelList;
  /** Get a Collection either by id or slug. If neither id nor slug is specified, an error will result. */
  collection?: Maybe<Collection>;
  collectionFilters: Array<ConfigurableOperationDefinition>;
  collections: CollectionList;
  countries: CountryList;
  country?: Maybe<Country>;
  customer?: Maybe<Customer>;
  customerGroup?: Maybe<CustomerGroup>;
  customerGroups: CustomerGroupList;
  customers: CustomerList;
  /** Returns a list of eligible shipping methods for the draft Order */
  eligibleShippingMethodsForDraftOrder: Array<ShippingMethodQuote>;
  emailTemplate: EmailTemplate;
  emailTemplateByCode: EmailTemplate;
  emailTemplates: Array<EmailTemplate>;
  facet?: Maybe<Facet>;
  facetValues: FacetValueList;
  facets: FacetList;
  fulfillmentHandlers: Array<ConfigurableOperationDefinition>;
  getOrderMessage?: Maybe<OrderMessage>;
  getOrderMessages: OrderMessageList;
  getUpsale: UpSale;
  globalSettings: GlobalSettings;
  job?: Maybe<Job>;
  jobBufferSize: Array<JobBufferSize>;
  jobQueues: Array<JobQueue>;
  jobs: JobList;
  jobsById: Array<Job>;
  listByProduct: UpSaleList;
  me?: Maybe<CurrentUser>;
  /** Get metrics for the given interval and metric types. */
  metricSummary: Array<MetricSummary>;
  order?: Maybe<Order>;
  orders: OrderList;
  page?: Maybe<Page>;
  pages: PageList;
  partials: Array<EmailTemplate>;
  paymentMethod?: Maybe<PaymentMethod>;
  paymentMethodEligibilityCheckers: Array<ConfigurableOperationDefinition>;
  paymentMethodHandlers: Array<ConfigurableOperationDefinition>;
  paymentMethods: PaymentMethodList;
  pendingSearchIndexUpdates: Scalars['Int']['output'];
  /** Used for real-time previews of the contents of a Collection */
  previewCollectionVariants: ProductVariantList;
  /** Get a Product either by id or slug. If neither id nor slug is specified, an error will result. */
  product?: Maybe<Product>;
  productAdditionalDetails: ProductType;
  productAdditionalDetailsForProduct: ProductType;
  productOptionGroup?: Maybe<ProductOptionGroup>;
  productOptionGroups: Array<ProductOptionGroup>;
  /** Get a ProductVariant by id */
  productVariant?: Maybe<ProductVariant>;
  /** List ProductVariants either all or for the specific product. */
  productVariants: ProductVariantList;
  /** List Products */
  products: ProductList;
  promotion?: Maybe<Promotion>;
  promotionActions: Array<ConfigurableOperationDefinition>;
  promotionConditions: Array<ConfigurableOperationDefinition>;
  promotions: PromotionList;
  province?: Maybe<Province>;
  provinces: ProvinceList;
  role?: Maybe<Role>;
  roles: RoleList;
  search: SearchResponse;
  seller?: Maybe<Seller>;
  sellers: SellerList;
  shippingCalculators: Array<ConfigurableOperationDefinition>;
  shippingEligibilityCheckers: Array<ConfigurableOperationDefinition>;
  shippingMethod?: Maybe<ShippingMethod>;
  shippingMethods: ShippingMethodList;
  stockLocation?: Maybe<StockLocation>;
  stockLocations: StockLocationList;
  tag: Tag;
  tags: TagList;
  taxCategories: TaxCategoryList;
  taxCategory?: Maybe<TaxCategory>;
  taxRate?: Maybe<TaxRate>;
  taxRates: TaxRateList;
  testEligibleShippingMethods: Array<ShippingMethodQuote>;
  testShippingMethod: TestShippingMethodResult;
  zone?: Maybe<Zone>;
  zones: ZoneList;
};
// ...
export type ProductDetailsQueryVariables = Exact<{
  id: Scalars['ID']['input'];
}>;


export type ProductDetailsQuery = { __typename?: 'Query', productAdditionalDetailsForProduct: { __typename?: 'ProductType', slogan: string, alternateName: string, additionalType: string, disambiguatingDescription: string, oneWordName: string, isService: boolean, serviceOutput?: string, serviceType?: string, repetitions: Array<number>, meta_title: string, meta_keyword: string, meta_description: string, productId: string } };

export type GetProductTypesForProductQueryVariables = Exact<{
  productId: Scalars['ID']['input'];
}>;


export type GetProductTypesForProductQuery = { __typename?: 'Query', productAdditionalDetailsForProduct: { __typename?: 'ProductType', slogan: string, alternateName: string, additionalType: string, disambiguatingDescription: string, oneWordName: string, isService: boolean, serviceOutput?: string, serviceType?: string, repetitions: Array<number>, meta_title: string, meta_keyword: string, meta_description: string } };

export type CreateProductDetailsMutationVariables = Exact<{
  input: CreateProductTypeInput;
}>;


export type CreateProductDetailsMutation = { __typename?: 'Mutation', createProductAdditionalDetails: { __typename?: 'ProductType', slogan: string, alternateName: string, additionalType: string, disambiguatingDescription: string, oneWordName: string, isService: boolean, serviceOutput?: string, serviceType?: string, repetitions: Array<number>, meta_title: string, meta_keyword: string, meta_description: string, productId: string } };
// ...
export const ProductDetailsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"ProductDetails"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ID"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"productAdditionalDetailsForProduct"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"productId"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"slogan"}},{"kind":"Field","name":{"kind":"Name","value":"alternateName"}},{"kind":"Field","name":{"kind":"Name","value":"additionalType"}},{"kind":"Field","name":{"kind":"Name","value":"disambiguatingDescription"}},{"kind":"Field","name":{"kind":"Name","value":"oneWordName"}},{"kind":"Field","name":{"kind":"Name","value":"isService"}},{"kind":"Field","name":{"kind":"Name","value":"serviceOutput"}},{"kind":"Field","name":{"kind":"Name","value":"serviceType"}},{"kind":"Field","name":{"kind":"Name","value":"repetitions"}},{"kind":"Field","name":{"kind":"Name","value":"meta_title"}},{"kind":"Field","name":{"kind":"Name","value":"meta_keyword"}},{"kind":"Field","name":{"kind":"Name","value":"meta_description"}},{"kind":"Field","name":{"kind":"Name","value":"productId"}}]}}]}}]} as unknown as DocumentNode<ProductDetailsQuery, ProductDetailsQueryVariables>;
// ...
export const CreateProductDetailsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"mutation","name":{"kind":"Name","value":"CreateProductDetails"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"input"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"CreateProductTypeInput"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"createProductAdditionalDetails"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"Variable","name":{"kind":"Name","value":"input"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"slogan"}},{"kind":"Field","name":{"kind":"Name","value":"alternateName"}},{"kind":"Field","name":{"kind":"Name","value":"additionalType"}},{"kind":"Field","name":{"kind":"Name","value":"disambiguatingDescription"}},{"kind":"Field","name":{"kind":"Name","value":"oneWordName"}},{"kind":"Field","name":{"kind":"Name","value":"isService"}},{"kind":"Field","name":{"kind":"Name","value":"serviceOutput"}},{"kind":"Field","name":{"kind":"Name","value":"serviceType"}},{"kind":"Field","name":{"kind":"Name","value":"repetitions"}},{"kind":"Field","name":{"kind":"Name","value":"meta_title"}},{"kind":"Field","name":{"kind":"Name","value":"meta_keyword"}},{"kind":"Field","name":{"kind":"Name","value":"meta_description"}},{"kind":"Field","name":{"kind":"Name","value":"productId"}}]}}]}}]} as unknown as DocumentNode<CreateProductDetailsMutation, CreateProductDetailsMutationVariables>;